ymir-sdl3/App: Move mouse capture to MouseCaptureService
