+++ apps/ymir-sdl3/CMakeLists.txt
@@ -75,6 +75,8 @@ add_executable(ymir-sdl3
     src/app/services/screenshot_types.hpp
     src/app/services/update_checker_service.cpp
     src/app/services/update_checker_service.hpp
+    src/app/services/mouse_capture_service.cpp
+    src/app/services/mouse_capture_service.hpp
 
     src/app/ui/window_base.cpp
     src/app/ui/window_base.hpp
