+++ apps/ymir-sdl3/src/app/app.cpp
@@ -161,6 +161,7 @@ App::App()
     : m_saveStateService()
     , m_midiService(m_context.serviceLocator)
     , m_settings(m_context)
+    , m_mouseCaptureService(m_context, m_settings)
     , m_systemStateWindow(m_context)
     , m_bupMgrWindow(m_context)
     , m_masterSH2WindowSet(m_context, true)
@@ -184,6 +185,7 @@ App::App()
     m_context.serviceLocator.Register(m_settings);
     m_context.serviceLocator.Register(m_screenshotService);
     m_context.serviceLocator.Register(m_updateCheckerService);
+    m_context.serviceLocator.Register(m_mouseCaptureService);
     m_settings.BindConfiguration(m_context.saturn.instance->configuration);
 
     // Preinitialize some memory viewers
@@ -228,15 +230,9 @@ int App::Run(const CommandLineOptions &options) {
         for (uint32 port = 0; port < 2; ++port) {
             inputSettings.ports[port].type.Observe([&, port = port](ymir::peripheral::PeripheralType type) {
                 m_context.EnqueueEvent(events::emu::InsertPeripheral(port, type));
-                bool changed;
-                if (type == ymir::peripheral::PeripheralType::VirtuaGun ||
-                    type == ymir::peripheral::PeripheralType::ShuttleMouse) {
-                    changed = m_validPeripheralsForMouseCapture.insert(port).second;
-                } else {
-                    changed = m_validPeripheralsForMouseCapture.erase(port) > 0;
-                }
+                bool changed = m_mouseCaptureService.SetPeripheralType(port, type);
                 if (changed) {
-                    ReleaseAllMice();
+                    m_mouseCaptureService.ReleaseAllMice();
                 }
             });
         }
@@ -1895,7 +1891,8 @@ void App::RunEmulator() {
         }
     }
 
-    settings.input.mouse.captureMode.ObserveAndNotify([&](Settings::Input::Mouse::CaptureMode) { ReleaseAllMice(); });
+    settings.input.mouse.captureMode.ObserveAndNotify(
+        [&](Settings::Input::Mouse::CaptureMode) { m_mouseCaptureService.ReleaseAllMice(); });
 
     RebindInputs();
 
@@ -2233,7 +2230,7 @@ void App::RunEmulator() {
                     }
 
                     // Restore system mouse cursor and release captured mice
-                    ReleaseAllMice();
+                    m_mouseCaptureService.ReleaseAllMice();
                 }
                 break;
 
@@ -2247,13 +2244,14 @@ void App::RunEmulator() {
                 if (evt.button.which != SDL_PEN_MOUSEID && evt.button.which != SDL_TOUCH_MOUSEID) {
                     inputContext.DisconnectMouse(evt.mdevice.which);
                     devlog::debug<grp::base>("Mouse {} removed", evt.mdevice.which);
-                    ReleaseMouse(evt.mdevice.which);
+                    m_mouseCaptureService.ReleaseMouse(evt.mdevice.which);
                 }
                 break;
             case SDL_EVENT_MOUSE_BUTTON_DOWN: [[fallthrough]];
             case SDL_EVENT_MOUSE_BUTTON_UP:
                 if (!io.WantCaptureMouse) {
-                    if (!IsMouseCaptured() && !HasValidPeripheralsForMouseCapture() &&
+                    if (!m_mouseCaptureService.IsMouseCaptured() &&
+                        !m_mouseCaptureService.HasValidPeripheralsForMouseCapture() &&
                         settings.video.doubleClickToFullScreen && evt.button.clicks % 2 == 0 && evt.button.down &&
                         evt.button.button == SDL_BUTTON_LEFT) {
                         settings.video.fullScreen = !settings.video.fullScreen;
@@ -2269,7 +2267,7 @@ void App::RunEmulator() {
 
                         // Try capturing the mouse cursor
                         if (evt.button.down && evt.button.button == SDL_BUTTON_LEFT) {
-                            ConnectMouseToPeripheral(evt.button.which);
+                            m_mouseCaptureService.ConnectMouseToPeripheral(evt.button.which);
                         }
                     }
                 }
@@ -2407,7 +2405,7 @@ void App::RunEmulator() {
                         m_context.EnqueueEvent(events::emu::SetPaused(true));
                     }
                 }
-                ReleaseAllMice();
+                m_mouseCaptureService.ReleaseAllMice();
                 break;
 
             case SDL_EVENT_DROP_FILE: //
@@ -2699,8 +2697,8 @@ void App::RunEmulator() {
         if (mouseMoved || mouseDown || io.WantCaptureMouse) {
             m_mouseHideTime = clk::now();
         }
-        const bool hideMouse = (m_systemMouseCaptured && !io.WantCaptureMouse) || !m_capturedMice.empty() ||
-                               clk::now() >= m_mouseHideTime + 2s;
+        const bool hideMouse =
+            m_mouseCaptureService.ShouldHideMouse(io.WantCaptureMouse) || clk::now() >= m_mouseHideTime + 2s;
         if (hideMouse) {
             ImGui::SetMouseCursor(ImGuiMouseCursor_None);
         }
@@ -2724,8 +2722,8 @@ void App::RunEmulator() {
         // NOTE: ImGui::IsWindowFocused(ImGuiFocusedFlags_AnyWindow) cannot be used becaused the application always
         // starts with main menu bar focused
         if (settings.input.mouse.captureMode == Settings::Input::Mouse::CaptureMode::PhysicalMouse &&
-            io.WantCaptureKeyboard && (!m_capturedMice.empty() || m_mouseCaptureActive)) {
-            ReleaseAllMice();
+            io.WantCaptureKeyboard && m_mouseCaptureService.IsPhysicalMouseCapturedOrActive()) {
+            m_mouseCaptureService.ReleaseAllMice();
         }
 
         auto *viewport = ImGui::GetMainViewport();
@@ -3364,7 +3362,7 @@ void App::RunEmulator() {
                     (void *)&aspectRatio);
 
                 ImGuiWindowFlags windowFlags = ImGuiWindowFlags_NoNavInputs;
-                if (m_systemMouseCaptured || HasValidPeripheralsForMouseCapture()) {
+                if (m_mouseCaptureService.ShouldDisableImGuiMouseInputs()) {
                     windowFlags |= ImGuiWindowFlags_NoMouseInputs;
                 }
                 if (ImGui::Begin(title.c_str(), &videoSettings.displayVideoOutputInWindow, windowFlags)) {
@@ -4301,259 +4299,6 @@ void App::DrawInputs(ImDrawList *drawList) {
             ui::widgets::Crosshair(drawList, params, {input.posX, input.posY});
         }
     }
-}
-
-bool App::CaptureMouse(uint32 id, uint32 port) {
-    // Prevent capturing global mouse
-    if (id == 0) {
-        return false;
-    }
-
-    // Bail out if mouse is already bound to a peripheral
-    if (m_capturedMice.contains(id)) {
-        return false;
-    }
-
-    // Bind mouse to peripheral
-    m_capturedMice[id] = port;
-    const char *name = SDL_GetMouseNameForID(id);
-    if (name != nullptr) {
-        m_context.DisplayMessage(fmt::format("{} bound to {}", name, GetPeripheralName(port)));
-    } else {
-        m_context.DisplayMessage(fmt::format("Mouse {} bound to {}", id, GetPeripheralName(port)));
-    }
-    ConfigureMouseCapture();
-
-    m_mouseCaptureActive = false;
-
-    return true;
-}
-
-bool App::ReleaseMouse(uint32 id) {
-    // Prevent releasing global mouse
-    if (id == 0) {
-        return false;
-    }
-
-    // Bail out if mouse not bound to a peripheral
-    if (!m_capturedMice.contains(id)) {
-        return false;
-    }
-
-    // Release mouse
-    const uint32 port = m_capturedMice.at(id);
-    m_capturedMice.erase(id);
-
-    const char *name = SDL_GetMouseNameForID(id);
-    if (name != nullptr) {
-        m_context.DisplayMessage(fmt::format("{} released from {}", name, GetPeripheralName(port)));
-    } else {
-        m_context.DisplayMessage(fmt::format("Mouse {} released from {}", id, GetPeripheralName(port)));
-    }
-
-    if (m_capturedMice.empty()) {
-        m_context.DisplayMessage("Leaving mouse capture mode");
-    }
-
-    ConfigureMouseCapture();
-    (void)m_context.inputContext.UnmapMouseInputs(id);
-
-    return true;
-}
-
-bool App::CaptureSystemMouse(uint32 port) {
-    const bool captured = !m_systemMouseCaptured;
-    if (captured) {
-        m_context.DisplayMessage(fmt::format("Mouse cursor bound to {}", GetPeripheralName(port)));
-        m_context.DisplayMessage("Press ESC to release");
-        ConfigureMouseCapture();
-        m_systemMouseCaptured = true;
-        m_systemMousePeripheral = port;
-    }
-    return captured;
-}
-
-bool App::ReleaseSystemMouse() {
-    const bool released = m_systemMouseCaptured;
-    if (released) {
-        m_context.DisplayMessage(
-            fmt::format("Mouse cursor released from {}", GetPeripheralName(m_systemMousePeripheral)));
-        ConfigureMouseCapture();
-        m_systemMouseCaptured = false;
-        m_context.virtuaGunInputs[m_systemMousePeripheral].mouseAbsolute = false;
-        (void)m_context.inputContext.UnmapMouseInputs(0);
-    }
-    return released;
-}
-
-void App::ReleaseAllMice() {
-    bool released = m_mouseCaptureActive || m_systemMouseCaptured || !m_capturedMice.empty();
-
-    if (m_systemMouseCaptured) {
-        m_systemMouseCaptured = false;
-        (void)m_context.inputContext.UnmapMouseInputs(0);
-    }
-    for (auto [id, _] : m_capturedMice) {
-        (void)m_context.inputContext.UnmapMouseInputs(id);
-    }
-    m_capturedMice.clear();
-
-    m_mouseCaptureActive = false;
-
-    for (uint32 i = 0; i < 2; ++i) {
-        m_context.virtuaGunInputs[i].mouseAbsolute = false;
-    }
-
-    if (released) {
-        m_context.DisplayMessage("All mice released");
-        ConfigureMouseCapture();
-    }
-}
-
-void App::ConfigureMouseCapture() {
-    const bool grabSystemCursor = false; // TODO: pull from settings
-    const bool relativeMode = m_mouseCaptureActive || !m_capturedMice.empty();
-    const bool captured = (m_systemMouseCaptured && grabSystemCursor) || relativeMode;
-    const bool grabbed = captured && !relativeMode;
-    const bool wasRelativeMode = SDL_GetWindowRelativeMouseMode(m_context.screen.window);
-    SDL_SetWindowMouseGrab(m_context.screen.window, grabbed);
-    SDL_SetWindowRelativeMouseMode(m_context.screen.window, relativeMode);
-    if (captured) {
-        ImGui::GetIO().ConfigFlags |= ImGuiConfigFlags_NoMouse;
-    } else {
-        ImGui::GetIO().ConfigFlags &= ~ImGuiConfigFlags_NoMouse;
-    }
-
-    if (wasRelativeMode != relativeMode) {
-        if (relativeMode) {
-            m_context.DisplayMessage("Entering mouse capture mode");
-            m_context.DisplayMessage("Press ESC to release mice");
-        } else {
-            m_context.DisplayMessage("Leaving mouse capture mode");
-        }
-    }
-
-    devlog::debug<grp::base>("Mouse capture config: captured={} relative={} grab={}", captured, relativeMode, grabbed);
-}
-
-void App::ConnectMouseToPeripheral(uint32 id) {
-    using PeripheralType = ymir::peripheral::PeripheralType;
-    using CaptureMode = Settings::Input::Mouse::CaptureMode;
-
-    // Build list of candidate peripherals.
-    // Bail out if none are available.
-    const std::set<uint32> candidates = GetCandidatePeripheralsForMouseCapture();
-    if (candidates.empty()) {
-        return;
-    }
-
-    const auto &settings = m_settings;
-    CaptureMode captureMode = settings.input.mouse.captureMode;
-
-    // Force PhysicalMouse mode if a Shuttle Mouse is connected to any port
-    for (uint32 port : candidates) {
-        if (settings.input.ports[port].type == PeripheralType::ShuttleMouse) {
-            captureMode = CaptureMode::PhysicalMouse;
-            break;
-        }
-    }
-
-    // Bail out if mouse is already bound to a peripheral
-    if (captureMode == CaptureMode::SystemCursor && m_systemMouseCaptured) {
-        return;
-    }
-    if (captureMode == CaptureMode::PhysicalMouse && m_capturedMice.contains(id)) {
-        return;
-    }
-
-    if (captureMode == CaptureMode::SystemCursor) {
-        // Force global mouse ID when in system cursor capture mode
-        id = 0;
-    } else if (id == 0) {
-        // Engage mouse capture in physical mouse capture mode when using the global mouse ID
-        m_mouseCaptureActive = true;
-        ConfigureMouseCapture();
-        return;
-    }
-
-    // TODO: allow user to pick port when there are multiple candidates
-    // - show a popup asking which one the player wants to bind to (or Cancel)
-    // - cancel popup:
-    //   - on ESC press
-    //   - if the app loses focus
-    //   - if mouse capture mode changes
-    //   - if the target mouse is removed
-    //   - if the valid controller list is changed
-    const uint32 port = *candidates.begin();
-    const PeripheralType type = settings.input.ports[port].type;
-    std::string periphName = GetPeripheralName(port);
-
-    assert(type == PeripheralType::VirtuaGun || type == PeripheralType::ShuttleMouse);
-
-    const bool captured = [&] {
-        switch (captureMode) {
-        case CaptureMode::SystemCursor: return CaptureSystemMouse(port);
-        case CaptureMode::PhysicalMouse: return CaptureMouse(id, port);
-        }
-        return false;
-    }();
-
-    // Bind inputs
-    auto &inputCtx = m_context.inputContext;
-    if (type == PeripheralType::VirtuaGun) {
-        auto *actionCtx = &m_context.virtuaGunInputs[port];
-        if (captureMode == CaptureMode::PhysicalMouse) {
-            actionCtx->mouseAbsolute = false;
-            (void)inputCtx.MapAction({id, input::MouseAxis2D::MouseRelative}, actions::virtua_gun::MouseRelMove,
-                                     actionCtx);
-        } else {
-            actionCtx->mouseAbsolute = true;
-            (void)inputCtx.MapAction({0, input::MouseAxis2D::MouseAbsolute}, actions::virtua_gun::MouseAbsMove,
-                                     actionCtx);
-        }
-        // TODO: map inputs according to settings
-        (void)inputCtx.MapAction({id, input::MouseButton::Left}, actions::virtua_gun::Trigger, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Right}, actions::virtua_gun::Reload, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Middle}, actions::virtua_gun::Start, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Extra1}, actions::virtua_gun::Start, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Extra2}, actions::virtua_gun::Start, actionCtx);
-    } else if (type == PeripheralType::ShuttleMouse) {
-        assert(captureMode == CaptureMode::PhysicalMouse);
-
-        auto *actionCtx = &m_context.shuttleMouseInputs[port];
-        (void)inputCtx.MapAction({id, input::MouseAxis2D::MouseRelative}, actions::shuttle_mouse::MouseRelMove,
-                                 actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Left}, actions::shuttle_mouse::Left, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Right}, actions::shuttle_mouse::Right, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Middle}, actions::shuttle_mouse::Middle, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Extra1}, actions::shuttle_mouse::Start, actionCtx);
-        (void)inputCtx.MapAction({id, input::MouseButton::Extra2}, actions::shuttle_mouse::Start, actionCtx);
-    }
-}
-
-bool App::IsMouseCaptured() const {
-    return m_systemMouseCaptured || !m_capturedMice.empty();
-}
-
-bool App::HasValidPeripheralsForMouseCapture() const {
-    return !GetCandidatePeripheralsForMouseCapture().empty();
-}
-
-std::set<uint32> App::GetCandidatePeripheralsForMouseCapture() const {
-    std::set<uint32> candidates = m_validPeripheralsForMouseCapture;
-    for (auto [_, port] : m_capturedMice) {
-        candidates.erase(port);
-    }
-    if (m_systemMouseCaptured) {
-        candidates.erase(m_systemMousePeripheral);
-    }
-    return candidates;
-}
-
-std::string App::GetPeripheralName(uint32 port) const {
-    const auto &settings = m_settings;
-    const ymir::peripheral::PeripheralType type = settings.input.ports[port].type;
-    return fmt::format("{} on port {}", ymir::peripheral::GetPeripheralName(type), port + 1);
 }
 
 std::pair<float, float> App::WindowToScreen(float x, float y) const {
